# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TinyCongress API",
    "description": "REST API for TinyCongress",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "REST API v1"
    }
  ],
  "paths": {
    "/build-info": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get build information",
        "description": "Returns metadata about the running service including version, git SHA, and build time.\n\n# Errors\n\nReturns `ProblemDetails` on internal server errors.",
        "operationId": "get_build_info",
        "responses": {
          "200": {
            "description": "Build information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildInfo"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BuildInfo": {
        "type": "object",
        "description": "Build metadata exposed via GraphQL, REST, and logs.",
        "required": [
          "version",
          "gitSha",
          "buildTime"
        ],
        "properties": {
          "buildTime": {
            "type": "string"
          },
          "gitSha": {
            "type": "string"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "version": {
            "type": "string"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details error response.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence"
          },
          "extensions": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProblemExtensions",
                "description": "Additional error details"
              }
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "URI reference identifying the specific occurrence"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary"
          },
          "type": {
            "type": "string",
            "description": "URI reference identifying the problem type"
          }
        }
      },
      "ProblemExtensions": {
        "type": "object",
        "description": "Extended error information mapping to GraphQL error codes.",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code matching GraphQL error codes"
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "Field that caused the error (for validation errors)"
          }
        }
      }
    }
  }
}